requirements.txt
Django==1.8
Pillow==2.5.3
argparse==1.2.1
distribute==0.6.24
django-chunks2==2.8.2
django-markdown==0.8.4
easy-thumbnails==2.1
wsgiref==0.1.2
psycopg2==2.5.4
dj-static==0.0.6
